requirements.txt
fastai==2.7.18
flask==2.3.3
gunicorn==21.2.0
pillow==10.0.0